Remove dom-node-iterator polyfill
The polyfill was previously used indirectly by `dom-seek` and directly
in `src/index.js` to polyfill the `NodeIterator.referenceNode` property
and optional-ness of `document.createNodeIterator` arguments.

Since there is no longer a dependency on `dom-seek` and the remaining
IE issues are easy to work around, we can simplify the overall
complexity by removing the polyfill.